TrapDoors More dependent on Room Size and size updates
So this is partially true. I ideally would like to have a solution to this dependent on the room sizes at a factor passed in to the trap. But for right now they are still hard coded because doing so would take far more work than time I have at this moment.