Trapdoors are now 100% dependent on rooms total size!
Please note, if the grounds of the rooms change or how low or high they are, these fascets DO NOT AUTOMATCIALLY CHANGE. There really isn't a way to do that without some slightly more complicated systems being in place throughout many more areas. Which would not be ideal if I'm being honest. This is the fastest and simplest solution so I this is the solutoin that was went with.